+++ Dockerfile.notification
@@ -6,13 +6,15 @@ WORKDIR /app
 
 ## 패키지 파일 복사 및 의존성 설치
 COPY package*.json ./
+COPY global.d.ts ./
+COPY nest-cli.json ./
 RUN npm install
 
 ## 소스 코드 복사
-COPY . .
+COPY notification/ notification/
 
 ## 애플리케이션 빌드
-RUN npm run build
+RUN npm run build:notification
 
 # 2. 실행 단계
 FROM node:18-alpine
