+++ nest-cli.json
@@ -1,8 +1,29 @@
 {
   "$schema": "https://json.schemastore.org/nest-cli",
   "collection": "@nestjs/schematics",
+  "monorepo": true,
   "sourceRoot": "src",
   "entryFile": "notification/main",
+  "projects": {
+    "notification": {
+      "type": "application",
+      "root": "notification",
+      "entryFile": "main",
+      "sourceRoot": "notification/src",
+      "compilerOptions": {
+        "tsConfigPath": "notification/tsconfig.build.json"
+      }
+    },
+    "worker": {
+      "type": "application",
+      "root": "worker",
+      "entryFile": "main",
+      "sourceRoot": "worker/src",
+      "compilerOptions": {
+        "tsConfigPath": "worker/tsconfig.json"
+      }
+    }
+  },
   "compilerOptions": {
     "deleteOutDir": true
   }
