+++ notification/src/domain/entity.ts
