+++ notification/src/infrastructure/auth/verifier.ts
