+++ notification/src/infrastructure/ormEntity.ts
@@ -1,6 +1,6 @@
 import { Column, Entity, PrimaryColumn } from "typeorm";
 
-import { NotificationStatus } from "@notification/application/dto";
+import { NotificationStatus } from "application/dto";
 
 @Entity("Notification")
 export default class NotificationORMEntity {
