+++ notification/src/presentation/dtos/ParametersDTO.ts
@@ -2,8 +2,8 @@ import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
 import { IsDate, IsOptional, ValidateIf, IsString } from "class-validator";
 
-import { NotificationStatus } from "@notification/domain/entity";
-import { IsTimeRange, AtLeastOneOption, TrimSeconds } from "@notification/utility/decorators";
+import { NotificationStatus } from "domain/entity";
+import { IsTimeRange, AtLeastOneOption, TrimSeconds } from "utility/decorators";
 
 export default class ParametersDTO {
     @IsDate()
