+++ notification/src/presentation/dtos/UpdateRequestDTO.ts
@@ -2,19 +2,19 @@ import { ApiProperty } from "@nestjs/swagger";
 import { Type } from "class-transformer";
 import { IsString, IsDate, IsOptional } from "class-validator";
 
-import { NotificationStatus } from "@notification/domain/entity";
-import { TrimSeconds } from "@notification/utility/decorators";
+import { NotificationStatus } from "domain/entity";
 
-export default class CreateRequestDTO {
+export default class UpdateRequestDTO {
     @IsString()
+    @IsOptional()
     @ApiProperty({ description: "수행될 이벤트 ID", example: "507f1f77bcf86cd799439011" })
     event_id: string;
 
     @IsDate()
     @Type(() => Date)
-    @TrimSeconds()
-    @ApiProperty({ description: "알림 발송 시간" })
-    send_at: Date;
+    @IsOptional()
+    @ApiProperty({ description: "알림 발송 시간", required: false })
+    send_at?: Date;
 
     @IsString()
     @IsOptional()
