+++ notification/src/utility/decorators/TrimSeconds.ts
