+++ package.json
@@ -6,13 +6,15 @@
   "private": true,
   "license": "UNLICENSED",
   "scripts": {
-    "build": "nest build",
+    "build": "npm run build:notification && npm run build:worker",
+    "build:notification": "nest build notification",
+    "build:worker": "nest build worker",
     "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-    "start": "nest start",
-    "start:dev": "nest start --watch",
-    "start:debug": "nest start --debug --watch",
-    "start:prod": "node dist/main",
-    "start:worker": "ts-node --require tsconfig-paths/register src/worker/main.ts",
+    "start": "nest start notification",
+    "start:dev": "nest start notification --watch",
+    "start:debug": "nest start notification --debug --watch",
+    "start:prod": "node dist/notification/main",
+    "start:worker": "nest start worker",
     "generate-openapi": "ts-node --require tsconfig-paths/register src/notification/utility/downloadOpenAPI.ts",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
