+++ worker/tsconfig.json
@@ -9,11 +9,8 @@
     "allowSyntheticDefaultImports": true,
     "target": "ES2021",
     "sourceMap": true,
-    "outDir": "./dist",
-    "baseUrl": "./",
-    "paths": {
-      "@notification/*": ["./src/notification/*"]
-    },
+    "outDir": "../dist/worker",
+    "baseUrl": "./src",
     "incremental": true,
     "skipLibCheck": true,
     "strictNullChecks": false,
@@ -22,6 +19,7 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false
   },
+  "include": ["src/**/*.ts", "../global.d.ts"],
   "ts-node": {
     "files": true
   }
